backend/backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers
PyJWT
pytz
psycopg2-binary==2.9.9
psycopg2
django-environ==0.11.2
drf-spectacular==0.27.2
drf_yasg
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
Pillow
django-extensions